processor support
The processor offers data cleanup, basic data validation, data persistence
and also handles autocomplete and other data related operations for the
form object.

The pixcore::form has been changed to accept a processor object on
creation; the processor will be used to autocomplete as well as check for
errors.

Various minior tweaks and improvements have been done to the existing
classes to account for the new functionality.

As part of the validation requirements for the Processor class a new
Validator class has been creating who is responsible for retrieving the
correct error messages.